+++ app/src/main/java/com/easefun/polyvsdk/PolyvApplication.java
@@ -14,7 +14,7 @@ public class PolyvApplication extends MultiDexApplication {
 
 	public static final String TAG = PolyvApplication.class.getSimpleName();
 
-	private static final String defaultConfig = "CMWht3MlpVkgpFzrLNAebYi4RdQDY/Nhvk3Kc+qWcck6chwHYKfl9o2aOVBvXVTRZD/14XFzVP7U5un43caq1FXwl0cYmTfimjTmNUYa1sZC1pkHE8gEsRpwpweQtEIiTGVEWrYVNo4/o5jI2/efzA==";
+	private static final String defaultConfig = "yQRmgnzPyCUYDx6weXRATIN8gkp7BYGAl3ATjE/jHZunrULx8CoKa1WGMjfHftVChhIQlCA9bFeDDX+ThiuBHLjsNRjotqxhiz97ZjYaCQH/MhUrbEURv58317PwPuGEf3rbLVPOa4c9jliBcO+22A==";
 
 	@Override
 	public void onCreate() {