+++ app/src/main/java/com/easefun/polyvsdk/player/PolyvPlayerMediaController.java
@@ -6,7 +6,6 @@ import android.app.PictureInPictureParams;
 import android.app.RemoteAction;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.ActivityInfo;
 import android.content.res.Configuration;
 import android.graphics.Bitmap;
 import android.graphics.Color;
@@ -296,7 +295,7 @@ public class PolyvPlayerMediaController extends PolyvBaseMediaController impleme
             canDragSeek = false;
         }
 
-        if (onDragSeekListener == null) {
+        if (onDragSeekListener == null || videoView == null) {
             return canDragSeek;
         }
         if (canDragSeek) {